+++ tests/hipscat_import/conftest.py
@@ -170,19 +170,19 @@ def resume_dir(test_data_dir):
 
 @pytest.fixture
 def small_sky_healsparse_map():
-    sub_map = healsparse.HealSparseMap.make_empty(hp.order2nside(0), hp.order2nside(0), np.int64, sentinel=0)
+    sub_map = healsparse.HealSparseMap.make_empty(hp.order2nside(0), hp.order2nside(0), np.int32, sentinel=0)
     sub_map[11] = 131
     return sub_map
 
 
 @pytest.fixture
 def empty_healsparse_order0():
-    return healsparse.HealSparseMap.make_empty(hp.order2nside(0), hp.order2nside(0), np.int64, sentinel=0)
+    return healsparse.HealSparseMap.make_empty(hp.order2nside(0), hp.order2nside(0), np.int32, sentinel=0)
 
 
 @pytest.fixture
 def empty_healsparse_order1():
-    return healsparse.HealSparseMap.make_empty(hp.order2nside(1), hp.order2nside(1), np.int64, sentinel=0)
+    return healsparse.HealSparseMap.make_empty(hp.order2nside(1), hp.order2nside(1), np.int32, sentinel=0)
 
 
 @pytest.fixture
